Fixes dialog bug in Alexa
When building the alexa interaction model there was a bug where even when empty the Dialog property
was still being included. This caused the builds to fail. This commit fixes that

There's new tests for that and also took the oportunity to refactor and fix some issues with the
path handling